requirements.txt
Django==1.3.1
autobreadcrumbs
django-crispy-forms
-e git+git://github.com/sveetch/django-braces.git#egg=django-braces
sveeaccounts
